Split KybSurveyService into projection, gate, and orchestration
Pure move: KybCheckFieldErrors owns the static check-to-field-error
projection with direct helper calls, OnboardingGate owns the blocked-
company decision and its audit event, and the survey service keeps the
orchestration; the moved test sections follow with synthetic personal
codes.

God classes 13 -> 12, PMD violations 152 -> 149.
