+++ metrics/scorecard.json
@@ -1,13 +1,13 @@
 {
-  "branchCoverage": 85.84,
-  "classCouplingViolations": 36,
+  "branchCoverage": 85.85,
+  "classCouplingViolations": 35,
   "clockViolations": 0,
   "cognitiveComplexityViolations": 20,
   "compilerWarnings": 0,
   "deepNestingViolations": 0,
   "disabledTests": 5,
   "disconnectedClasses": 454,
-  "godClasses": 13,
+  "godClasses": 12,
   "lineCoverage": 82.86,
   "longClasses": 4,
   "meanInstability": 0.473,
@@ -18,17 +18,17 @@
   "mutationsSliced": 1283,
   "mutationsTotal": 12781,
   "pitestSlicedModules": 6,
-  "pmdViolations": 152,
+  "pmdViolations": 149,
   "pmdViolationsByRule": {
     "CognitiveComplexity": 20,
-    "CouplingBetweenObjects": 36,
-    "CyclomaticComplexity": 78,
-    "GodClass": 13,
+    "CouplingBetweenObjects": 35,
+    "CyclomaticComplexity": 77,
+    "GodClass": 12,
     "TooManyMethods": 5
   },
-  "slowest10TestClassesSeconds": 111.2,
+  "slowest10TestClassesSeconds": 109.3,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 168,
+  "suiteWallTimeSeconds": 162,
   "testClassesOver10Seconds": 5,
   "top15ClassLines": 7092,
   "totalPackages": 227,
