+++ src/main/java/ee/tuleva/onboarding/kyb/survey/KybSurveyService.java
@@ -1,29 +1,19 @@
 package ee.tuleva.onboarding.kyb.survey;
 
 import static ee.tuleva.onboarding.event.TrackableEventType.SAVINGS_FUND_ONBOARDING_STATUS_CHANGE;
-import static ee.tuleva.onboarding.kyb.KybCheckType.DATA_CHANGED;
 import static ee.tuleva.onboarding.kyb.KybScreeningTrigger.SUBMISSION;
-import static ee.tuleva.onboarding.kyb.survey.BlockedReason.ALREADY_ONBOARDED;
 import static ee.tuleva.onboarding.kyb.survey.BlockedReason.NOT_BOARD_MEMBER;
-import static ee.tuleva.onboarding.kyb.survey.BlockedReason.ONBOARDING_PENDING;
-import static java.util.function.Function.identity;
-import static java.util.stream.Collectors.groupingBy;
 import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.mapping;
-import static java.util.stream.Collectors.toList;
-import static java.util.stream.Collectors.toMap;
 
 import ee.tuleva.onboarding.ariregister.CompanyDetail;
 import ee.tuleva.onboarding.ariregister.CompanyRelationship;
 import ee.tuleva.onboarding.event.TrackableSystemEvent;
 import ee.tuleva.onboarding.kyb.*;
-import ee.tuleva.onboarding.kyb.CompanyOnboarding;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.Nullable;
@@ -38,113 +28,10 @@ class KybSurveyService {
   private final LegalEntityScreener legalEntityScreener;
   private final KybSurveyResponseMapper kybSurveyResponseMapper;
   private final KybSurveyRepository kybSurveyRepository;
-  private final CompanyOnboarding companyOnboarding;
+  private final OnboardingGate onboardingGate;
   private final ApplicationEventPublisher eventPublisher;
 
-  private record FieldError(String field, ValidationError error) {}
-
-  private static final String KYC_MESSAGE = "Isikusamasuse tuvastamine on lõpetamata";
-  private static final String USER_KYC_MESSAGE = "Sinu isikusamasuse tuvastamine on lõpetamata";
-
-  // Projects a failed check to its client-facing field error(s). Each arm assigns a curated,
-  // client-safe code; the internal KybCheckType never reaches the wire. SANCTION and PEP collapse
-  // to one opaque code+message so a sanctions hit cannot be told apart from a PEP flag.
-  // RELATED_PERSONS_KYC needs runtime context (which persons are incomplete, and whether one is the
-  // onboarding user) and can yield two codes, so it is delegated to its own method.
-  private static Stream<FieldError> fieldErrorsFor(
-      KybCheck check, String userPersonalCode, List<RelatedPersonData> relatedPersons) {
-    return switch (check.type()) {
-      case COMPANY_ACTIVE ->
-          Stream.of(fieldError("status", "COMPANY_ACTIVE", "Ettevõte ei ole aktiivne"));
-      case HIGH_RISK_NACE ->
-          Stream.of(fieldError("naceCode", "UNSUPPORTED_NACE", "See tegevusala ei ole toetatud"));
-      case COMPANY_SANCTION, COMPANY_PEP ->
-          Stream.of(fieldError("name", "UNSERVICEABLE", "Ettevõtet ei ole võimalik teenindada"));
-      case RELATED_PERSONS_KYC -> relatedPersonsKycErrors(check, userPersonalCode, relatedPersons);
-      case COMPANY_LEGAL_FORM ->
-          Stream.of(fieldError("legalForm", "COMPANY_LEGAL_FORM", "Ainult OÜ on toetatud"));
-      case COMPANY_REGISTERED_IN_ESTONIA ->
-          Stream.of(
-              fieldError(
-                  "address",
-                  "COMPANY_REGISTERED_IN_ESTONIA",
-                  "Ettevõte ei ole registreeritud Eestis"));
-      case COMPANY_STRUCTURE,
-          SOLE_MEMBER_OWNERSHIP,
-          DUAL_MEMBER_OWNERSHIP,
-          SINGLE_BOARD_MEMBER_OWNERSHIP ->
-          Stream.of(
-              fieldError(
-                  "relatedPersons",
-                  "COMPANY_STRUCTURE",
-                  "Ettevõtte omandistruktuur ei ole toetatud"));
-      case DATA_CHANGED, SELF_CERTIFICATION, COMPANY_AGE -> Stream.of();
-    };
-  }
-
-  private static FieldError fieldError(String field, String code, String message) {
-    return new FieldError(field, new ValidationError(code, message));
-  }
-
-  private static FieldError fieldError(
-      String field, String code, String message, List<RelatedPersonData> persons) {
-    return new FieldError(field, new ValidationError(code, message, persons));
-  }
-
-  private static Stream<FieldError> relatedPersonsKycErrors(
-      KybCheck check, String userPersonalCode, List<RelatedPersonData> relatedPersons) {
-    var incompletePersonalCodes = RelatedPersonsKycMetadata.incompletePersonalCodes(check);
-    var userIncomplete = incompletePersonalCodes.contains(userPersonalCode);
-    var otherPersonalCodes =
-        incompletePersonalCodes.stream().filter(code -> !code.equals(userPersonalCode)).toList();
-
-    var errors = new ArrayList<FieldError>();
-    if (userIncomplete) {
-      errors.add(fieldError("relatedPersons", "USER_KYC", USER_KYC_MESSAGE));
-    }
-    if (!otherPersonalCodes.isEmpty()) {
-      errors.add(
-          fieldError(
-              "relatedPersons",
-              "OTHER_RELATED_PERSONS_KYC",
-              KYC_MESSAGE,
-              personsFor(otherPersonalCodes, relatedPersons)));
-    }
-    if (errors.isEmpty()) {
-      errors.add(fieldError("relatedPersons", "OTHER_RELATED_PERSONS_KYC", KYC_MESSAGE));
-    }
-    return errors.stream();
-  }
-
-  private static List<RelatedPersonData> personsFor(
-      List<String> personalCodes, List<RelatedPersonData> relatedPersons) {
-    var personsByPersonalCode =
-        relatedPersons.stream()
-            .filter(person -> person.personalCode() != null)
-            .collect(toMap(RelatedPersonData::personalCode, identity(), (a, b) -> a));
-    return personalCodes.stream()
-        .distinct()
-        .map(
-            personalCode ->
-                personsByPersonalCode.getOrDefault(
-                    personalCode, new RelatedPersonData(personalCode, null)))
-        .toList();
-  }
-
-  private static ValidationError blockedReasonError(BlockedReason reason) {
-    return new ValidationError(reason.name(), blockedReasonMessage(reason));
-  }
-
-  private static String blockedReasonMessage(BlockedReason reason) {
-    return switch (reason) {
-      case ALREADY_ONBOARDED -> "Ettevõte on juba liitunud";
-      case ONBOARDING_PENDING -> "Ettevõtte liitumine on pooleli";
-      case NOT_BOARD_MEMBER -> "Isik ei ole ettevõtte juhatuse liige";
-    };
-  }
-
   private static final String BOARD_MEMBER_ROLE = "JUHL";
-  private static final String LEGAL_ENTITY_PARTY_TYPE = "LEGAL_ENTITY";
 
   LegalEntityData initialValidation(String registryCode, String personalCode) {
     log.info(
@@ -156,7 +43,7 @@ class KybSurveyService {
     try {
       relationships = fetchAndVerifyBoardMember(registryCode, personalCode);
     } catch (NotBoardMemberException e) {
-      auditBlocked(registryCode, personalCode, NOT_BOARD_MEMBER);
+      onboardingGate.auditBlocked(registryCode, personalCode, NOT_BOARD_MEMBER);
       throw e;
     }
 
@@ -171,7 +58,7 @@ class KybSurveyService {
         relationships,
         result.checks(),
         personalCode,
-        getOnboardingError(registryCode));
+        onboardingGate.getOnboardingError(registryCode));
   }
 
   void submit(
@@ -184,16 +71,11 @@ class KybSurveyService {
     try {
       relationships = fetchAndVerifyBoardMember(registryCode, personalCode);
     } catch (NotBoardMemberException e) {
-      auditBlocked(registryCode, personalCode, NOT_BOARD_MEMBER);
+      onboardingGate.auditBlocked(registryCode, personalCode, NOT_BOARD_MEMBER);
       throw e;
     }
 
-    try {
-      verifyOnboardingAllowed(registryCode);
-    } catch (OnboardingNotAllowedException e) {
-      auditBlocked(registryCode, personalCode, e.getReason());
-      throw e;
-    }
+    onboardingGate.verifyOnboardingAllowed(registryCode, personalCode);
 
     // LatestKybSurveyInputs reads the newest survey for a company, so a rejected
     // submission must not leave one behind.
@@ -227,48 +109,6 @@ class KybSurveyService {
     return relationships;
   }
 
-  private Optional<BlockedReason> getBlockedReason(String registryCode) {
-    return companyOnboarding.findState(registryCode).flatMap(KybSurveyService::blockedReasonFor);
-  }
-
-  private static Optional<BlockedReason> blockedReasonFor(CompanyOnboarding.State status) {
-    return switch (status) {
-      case COMPLETED -> Optional.of(ALREADY_ONBOARDED);
-      case PENDING -> Optional.of(ONBOARDING_PENDING);
-      case REJECTED -> Optional.empty();
-    };
-  }
-
-  private Optional<ValidationError> getOnboardingError(String registryCode) {
-    return getBlockedReason(registryCode).map(KybSurveyService::blockedReasonError);
-  }
-
-  private void verifyOnboardingAllowed(String registryCode) {
-    getBlockedReason(registryCode)
-        .ifPresent(
-            reason -> {
-              throw new OnboardingNotAllowedException(registryCode, reason);
-            });
-  }
-
-  private void auditBlocked(String registryCode, String personalCode, BlockedReason reason) {
-    log.warn(
-        "Company onboarding blocked: registryCode={}, personalCode={}, reason={}",
-        registryCode,
-        personalCode,
-        reason);
-
-    var data = new LinkedHashMap<String, @Nullable Object>();
-    data.put("partyType", LEGAL_ENTITY_PARTY_TYPE);
-    data.put("registryCode", registryCode);
-    data.put("personalCode", personalCode);
-    data.put("outcome", "BLOCKED");
-    data.put("blockedReason", reason.name());
-
-    eventPublisher.publishEvent(
-        new TrackableSystemEvent(SAVINGS_FUND_ONBOARDING_STATUS_CHANGE, data));
-  }
-
   private void auditValidationFailures(
       String registryCode, String personalCode, List<KybCheck> checks) {
     if (checks.stream().allMatch(c -> !c.type().isOnboardingGate() || c.success())) {
@@ -285,7 +125,7 @@ class KybSurveyService {
             .collect(joining(",")));
 
     var data = new LinkedHashMap<String, @Nullable Object>();
-    data.put("partyType", LEGAL_ENTITY_PARTY_TYPE);
+    data.put("partyType", "LEGAL_ENTITY");
     data.put("registryCode", registryCode);
     data.put("personalCode", personalCode);
     data.put("outcome", "VALIDATION_FAILED");
@@ -311,9 +151,10 @@ class KybSurveyService {
       String userPersonalCode,
       Optional<ValidationError> onboardingError) {
 
-    var relatedPersons = dedupedByPersonalCode(relationships);
+    var relatedPersons = KybCheckFieldErrors.dedupedByPersonalCode(relationships);
 
-    var errorsByField = collectErrorsByField(checks, userPersonalCode, relatedPersons);
+    var errorsByField =
+        KybCheckFieldErrors.collectErrorsByField(checks, userPersonalCode, relatedPersons);
 
     var nameErrors = new ArrayList<>(errorsByField.getOrDefault("name", List.of()));
     onboardingError.ifPresent(nameErrors::add);
@@ -326,51 +167,19 @@ class KybSurveyService {
             .orElse(null);
 
     return new LegalEntityData(
-        validatedField(detail.getName(), nameErrors),
+        KybCheckFieldErrors.validatedField(detail.getName(), nameErrors),
         ValidatedField.valid(detail.getRegistryCode()),
-        validatedField(
+        KybCheckFieldErrors.validatedField(
             detail.getLegalForm().orElse(null), errorsByField.getOrDefault("legalForm", List.of())),
         ValidatedField.valid(detail.getFoundingDate().orElse(null)),
-        validatedField(status, errorsByField.getOrDefault("status", List.of())),
-        validatedField(
+        KybCheckFieldErrors.validatedField(status, errorsByField.getOrDefault("status", List.of())),
+        KybCheckFieldErrors.validatedField(
             LegalEntityAddress.fromCompanyAddress(detail.getAddress().orElse(null)),
             errorsByField.getOrDefault("address", List.of())),
         ValidatedField.valid(detail.getMainActivity().orElse(null)),
-        validatedField(
+        KybCheckFieldErrors.validatedField(
             detail.getNaceCode().orElse(null), errorsByField.getOrDefault("naceCode", List.of())),
-        validatedField(relatedPersons, errorsByField.getOrDefault("relatedPersons", List.of())));
-  }
-
-  private static List<RelatedPersonData> dedupedByPersonalCode(
-      List<CompanyRelationship> relationships) {
-    return relationships.stream()
-        .map(r -> new RelatedPersonData(r.personalCode(), formatName(r)))
-        .collect(
-            toMap(RelatedPersonData::personalCode, identity(), (a, b) -> a, LinkedHashMap::new))
-        .values()
-        .stream()
-        .toList();
-  }
-
-  private static Map<String, List<ValidationError>> collectErrorsByField(
-      List<KybCheck> checks, String userPersonalCode, List<RelatedPersonData> relatedPersons) {
-    return checks.stream()
-        .filter(check -> !check.success())
-        .flatMap(check -> fieldErrorsFor(check, userPersonalCode, relatedPersons))
-        .collect(groupingBy(FieldError::field, mapping(FieldError::error, toList())));
-  }
-
-  private static <T> ValidatedField<T> validatedField(
-      @Nullable T value, List<ValidationError> errors) {
-    return errors.isEmpty()
-        ? ValidatedField.valid(value)
-        : ValidatedField.withErrors(value, errors);
-  }
-
-  private static String formatName(CompanyRelationship r) {
-    if (r.firstName() != null && r.lastName() != null) {
-      return r.firstName() + " " + r.lastName();
-    }
-    return r.lastName();
+        KybCheckFieldErrors.validatedField(
+            relatedPersons, errorsByField.getOrDefault("relatedPersons", List.of())));
   }
 }
